Plumb in applicationProfiler to child/worker/reporter system
